feat(user,point): 유저, 포인트

+++ apps/commerce-api/src/main/java/com/loopers/interfaces/api/point/UserPointV1Dto.java
@@ -4,7 +4,7 @@ import com.loopers.application.point.UserPointInfo;
 
 public class UserPointV1Dto {
 
-    public record PointResponse(Integer point
+    public record PointResponse(Long point
     ) {
         public static PointResponse from(UserPointInfo info) {
             return new PointResponse(
@@ -13,7 +13,7 @@ public class UserPointV1Dto {
         }
     }
 
-    public record PointChargeRequest(Integer amount
+    public record PointChargeRequest(Long amount
     ) {
 
     }
